requirements.txt
# Core dependencies
colorama>=0.4.6

# LLM dependencies
langchain>=0.1.0
langchain-groq>=0.1.0

# Environment management
python-dotenv>=1.0.0
